requirements.txt
certifi==2018.4.16
diff-match-patch==20121119
Django==2.0.5
django-adaptors==0.2.5
django-bootstrap3==10.0.1
django-filter==1.1.0
django-import-export==1.0.1
django-tables2==1.21.2
djangorestframework==3.8.2
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.4
psycopg2==2.7.5
pytz==2018.5
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
